Add files via upload
Includes additional features to see if the accuracy can be improved. 
Added previous release velocity and the x and z release position as well as the current velocity and x and z position. 
Weights can be toggled on and off to see if that has any affect on the outcomes (toggled off only predicts fastballs)